requirements.txt
typing_extensions
discord.py
python-dotenv
beautifulsoup4
requests
watchdog
peewee
pymysql